requirements.txt
css_inline==0.14.6
Markdown==3.7
python-dotenv==1.0.1
Requests==2.32.3
selenium==4.29.0
google-genai==1.4.0